requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
alembic==1.13.2
jinja2==3.1.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic-settings>=2.1.0,<3.0.0
pydantic[email]>=2.0.0,<3.0.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
itsdangerous==2.2.0
email-validator>=2.0.0